Restructure checkIfLemmyInstance
